+++ contracts-modularized/ERC5727/ERC5727EnumerableUpgradeableDS.sol
@@ -1,26 +1,53 @@
 //SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.0;
 
-import "./ERC5727Core.sol";
+import "./ERC5727UpgradeableDS.sol";
 import "./interfaces/IERC5727EnumerableUpgradeable.sol";
-import "../ERC3525/ERC3525SlotEnumerableCore.sol";
+import "../ERC3525/ERC3525SlotEnumerableUpgradeableDS.sol";
 import "./ERC5727EnumerableStorage.sol";
 
 contract ERC5727EnumerableUpgradeableDS is
-    ERC5727Core,
-    ERC3525SlotEnumerableCore
+    IERC5727EnumerableUpgradeable,
+    ERC5727UpgradeableDS,
+    ERC3525SlotEnumerableUpgradeableDS
 {
     using EnumerableSetUpgradeable for EnumerableSetUpgradeable.UintSet;
     using EnumerableSetUpgradeable for EnumerableSetUpgradeable.AddressSet;
     using EnumerableMapUpgradeable for EnumerableMapUpgradeable.AddressToUintMap;
 
-    function __ERC5727Enumerable_init() internal onlyInitializing {
+    function init(
+        string memory name_,
+        string memory symbol_,
+        address admin_,
+        string memory baseUri_,
+        string memory version_
+    ) external virtual override initializer {
+        __EIP712_init_unchained(name_, version_);
+        __ERC721_init_unchained(name_, symbol_, baseUri_);
+        __ERC3525_init_unchained(18);
+        __ERC5727_init_unchained(admin_);
         __ERC5727Enumerable_init_unchained();
     }
 
     function __ERC5727Enumerable_init_unchained() internal onlyInitializing {}
 
-    function slotCountOfOwner(address owner) external view returns (uint256) {
+    function supportsInterface(
+        bytes4 interfaceId
+    )
+        public
+        view
+        virtual
+        override(IERC165Upgradeable, ERC3525UpgradeableDS, ERC5727UpgradeableDS)
+        returns (bool)
+    {
+        return
+            interfaceId == type(IERC5727EnumerableUpgradeable).interfaceId ||
+            super.supportsInterface(interfaceId);
+    }
+
+    function slotCountOfOwner(
+        address owner
+    ) external view override returns (uint256) {
         if (owner == address(0)) revert NullValue();
 
         return LibERC5727EnumerableStorage.s()._slotsOfOwner[owner].length();
@@ -29,7 +56,7 @@ contract ERC5727EnumerableUpgradeableDS is
     function slotOfOwnerByIndex(
         address owner,
         uint256 index
-    ) external view returns (uint256) {
+    ) external view override returns (uint256) {
         if (owner == address(0)) revert NullValue();
         uint256 slotCountByOwner = LibERC5727EnumerableStorage
             .s()
@@ -44,7 +71,7 @@ contract ERC5727EnumerableUpgradeableDS is
     function ownerBalanceInSlot(
         address owner,
         uint256 slot
-    ) public view returns (uint256) {
+    ) public view override returns (uint256) {
         if (owner == address(0)) revert NullValue();
         if (!_slotExists(slot)) revert NotFound(slot);
 
@@ -95,14 +122,12 @@ contract ERC5727EnumerableUpgradeableDS is
         address to,
         uint256 firstTokenId,
         uint256 batchSize
-    ) internal virtual override(ERC721EnumerableCore, ERC5727Core) {
+    )
+        internal
+        virtual
+        override(ERC721EnumerableUpgradeable, ERC5727UpgradeableDS)
+    {
         super._beforeTokenTransfer(from, to, firstTokenId, batchSize);
-    }
-
-    function _burn(
-        uint256 tokenId
-    ) internal virtual override(ERC3525Core, ERC5727Core) {
-        ERC5727Core._burn(tokenId);
     }
 
     function _beforeValueTransfer(
@@ -112,8 +137,21 @@ contract ERC5727EnumerableUpgradeableDS is
         uint256 toTokenId,
         uint256 slot,
         uint256 value
-    ) internal virtual override(ERC3525SlotEnumerableCore, ERC5727Core) {
-        super._beforeValueTransfer(
+    )
+        internal
+        virtual
+        override(ERC3525SlotEnumerableUpgradeableDS, ERC5727UpgradeableDS)
+    {
+        ERC3525SlotEnumerableUpgradeableDS._beforeValueTransfer(
+            from,
+            to,
+            fromTokenId,
+            toTokenId,
+            slot,
+            value
+        );
+
+        ERC5727UpgradeableDS._beforeValueTransfer(
             from,
             to,
             fromTokenId,
@@ -138,8 +176,12 @@ contract ERC5727EnumerableUpgradeableDS is
         uint256 toTokenId,
         uint256 slot,
         uint256 value
-    ) internal virtual override(ERC3525SlotEnumerableCore, ERC3525Core) {
-        super._afterValueTransfer(
+    )
+        internal
+        virtual
+        override(ERC3525SlotEnumerableUpgradeableDS, ERC3525UpgradeableDS)
+    {
+        ERC3525SlotEnumerableUpgradeableDS._afterValueTransfer(
             from,
             to,
             fromTokenId,