+++ scripts/deployModularized.ts
@@ -1,10 +1,9 @@
 import { ethers } from 'hardhat';
 import {
   type DiamondMultiInit__factory,
-  type SoulHubModularized,
   type TransparentUpgradeableProxy,
-  ERC5727UpgradeableDS__factory,
-  ERC5727GovernanceUpgradeableDS__factory,
+  SoulHubUpgradeable__factory,
+  type SoulHubUpgradeable,
 } from '../typechain';
 import { FacetCutAction, getSelectors, remove } from '../test/utils';
 
@@ -26,8 +25,8 @@ async function deployTransparentProxy(
   return transparentProxy;
 }
 
-async function deploySoulHub(): Promise<SoulHubModularized> {
-  const soulHubContract = await ethers.getContractFactory('SoulHubModularized');
+async function deploySoulHub(): Promise<SoulHubUpgradeable> {
+  const soulHubContract = await ethers.getContractFactory('SoulHubUpgradeable');
   const soulHub = await soulHubContract.deploy();
   await soulHub.deployed();
   console.log('SoulHub contract deployed to:', soulHub.address);
@@ -38,15 +37,15 @@ async function deployFacets(): Promise<FacetCuts[]> {
   const FacetNames = [
     'DiamondCutFacet',
     'DiamondLoupeFacet',
-    'ERC5727UpgradeableDS',
+    'ERC5727SlotSettableUpgradeableDS',
     'ERC5727ClaimableUpgradeableDS',
     'ERC5727GovernanceUpgradeableDS',
     'ERC5727RecoveryUpgradeableDS',
     'ERC5727ExpirableUpgradeableDS',
-    'ERC5727EnumerableUpgradeableDS',
     'ERC5727DelegateUpgradeableDS',
   ];
   // The `facetCuts` variable is the FacetCut[] that contains the functions to add during diamond deployment
+  let allSelectors: string[] = [];
   const facetCuts: FacetCuts[] = [];
   for (const FacetName of FacetNames) {
     const Facet = await ethers.getContractFactory(FacetName);
@@ -59,21 +58,20 @@ async function deployFacets(): Promise<FacetCuts[]> {
         action: FacetCutAction.Add,
         functionSelectors: getSelectors(facet),
       });
+      allSelectors = [...getSelectors(facet)];
     } else {
       facetCuts.push({
         facetAddress: facet.address,
         action: FacetCutAction.Add,
-        functionSelectors: remove(
-          getSelectors(facet),
-          facetCuts[facetCuts.length - 1].functionSelectors,
-        ),
+        functionSelectors: remove(getSelectors(facet), allSelectors),
       });
+      allSelectors = [...allSelectors, ...facetCuts[facetCuts.length - 1].functionSelectors];
     }
   }
   return facetCuts;
 }
 async function main(): Promise<void> {
-  const [admin] = await ethers.getSigners();
+  const [admin, owner] = await ethers.getSigners();
   console.log('Deploying contracts with the account:', admin.address);
   console.log('Network:', (await ethers.provider.getNetwork()).name);
   const facetCuts = await deployFacets();
@@ -81,40 +79,25 @@ async function main(): Promise<void> {
     'DiamondMultiInit',
   );
   const diamondMultiInit = await diamondMultiInitFactory.deploy();
+  await diamondMultiInit.deployed();
   console.log('diamondInit deployed: ', diamondMultiInit.address);
-  const IERC5727 = ERC5727UpgradeableDS__factory.createInterface();
-  const IERC5727Governance = ERC5727GovernanceUpgradeableDS__factory.createInterface();
-  const ERC5727InitCall = IERC5727.encodeFunctionData('init', [
-    'soulhub',
-    'SOUL',
-    admin.address,
-    '1',
-  ]);
-  const ERC5727GovernanceInitCall = IERC5727Governance.encodeFunctionData('init', [admin.address]);
-  // Creating a function call
-  // This call gets executed during deployment and can also be executed in upgrades
-  // It is executed with delegatecall on the DiamondInit address.
-  const abi = [
-    'function multiInit(address[] calldata _addresses, bytes[] calldata _calldata) external',
-  ];
-  const diamondMultiInitInterface = new ethers.utils.Interface(abi);
-  const functionCall = diamondMultiInitInterface.encodeFunctionData('multiInit', [
-    [facetCuts[2].facetAddress, facetCuts[4].facetAddress],
-    [ERC5727InitCall, ERC5727GovernanceInitCall],
-  ]);
-  // Setting arguments that will be used in the diamond constructor
-  const diamondArgs = {
-    owner: admin.address,
-    init: diamondMultiInit.address,
-    initCalldata: functionCall,
-  };
-  const Diamond = await ethers.getContractFactory('Diamond');
-  const diamond = await Diamond.deploy(facetCuts, diamondArgs);
-  await diamond.deployed();
+  const facetAddress = facetCuts.map((facet) => facet.facetAddress);
+  const action = facetCuts.map((facet) => facet.action);
+  const functionSelectors = facetCuts.map((facet) => facet.functionSelectors);
+  const initData = SoulHubUpgradeable__factory.createInterface().encodeFunctionData(
+    '__SoulHub_init',
+    [facetCuts.length, facetAddress, action, functionSelectors, diamondMultiInit.address],
+  );
   const soulHubImpl = await deploySoulHub();
-  const soulHubProxy = await deployTransparentProxy(soulHubImpl.address, admin.address, '0x');
+  await soulHubImpl.deployed();
+  console.log('soulHubImpl deployed', soulHubImpl.address);
+  const soulHubProxy = await deployTransparentProxy(soulHubImpl.address, admin.address, initData);
+
   console.log('SoulHubProxy deployed: ', soulHubProxy.address);
-  console.log('Diamond deployed: ', diamond.address);
+  const soulHubProxyContract = SoulHubUpgradeable__factory.connect(soulHubProxy.address, owner);
+  const tx = await soulHubProxyContract.createOrganization('1q23');
+  await tx.wait();
+  console.log(tx);
 }
 
 // We recommend this pattern to be able to use async/await everywhere